Remove `read2`, plus some other changes
`read2` was largely redundant with `read`, but it implemented a way for
the FromStr impl to return a "skippable" error, which makes `read_input`
ignore that line. I worked out a way to achieve the same thing in a
backward-compatible way for `read`. A FromStr impl can optionally use
`read::ParseErr` as its Err type, which allows it to signal that an
input line can be skipped (i.e., ignored).

Also:

* implement mdist_to() for Coord3D
* implement + and += for Coord2D and Turn
* Grid: implement an alternate format method `format_str` which has a
  formatter Fn argument
* change `read` methods to only trim CR or LF chars from the end of the
  input line (rather than all whitespace)